refactor : LottoGenerator 분리
 1. Auto와 Manual을 분리
 2. Manual에 대한 테스트 코드 작성
