+++ src/main/java/lottogame/domain/Lotto.java
@@ -4,6 +4,8 @@ import java.util.Collections;
 import java.util.List;
 
 public class Lotto {
+    public static final int LOTTO_SIZE = 6;
+
     private List<LottoNumber> lotto;
 
     Lotto(List<LottoNumber> lotto) {
